Changed list items to have a unique id
to set the cursor after sorting properly.
The Set- and Get-Equals Methode of the list still are usefull,
if the user wants to get the index of a item.
The unique id within the example items are left in,
only to show how to make unique items with a proper equals function.