Changed Method return types from error to tea.Cmd
to signal, errors and changes within the list Model, to the caller of the Methods.
Hide (made private) Sort interface satisfying Methods because Swap would allow
silent(!) change of the list Model, this means since the Swap Method is not allowed to
return a value of type tea.Cmd we can't return a command to signify
a change of the cursor item, other changes or a (index) error.

- changed Methods to fail early and hard, so to not change the Model in
  any way if the request (-ed index) is wrong.

- rewritten top level commends.

- removed filling of the width and hight of the Lines method.
  Should be a bubbles function and not a single part of the list-bubble.

- added list-model message structs to signify a change within the list Model.

- changed/removed tests accordingly.