changed Suf- and Prefixer interfaces
to be able to change the suf- and prefix width on each item,
rather than to force a unified width for suffix and prefix per draw.

This allows for different width for each item (not line),
with the expense of performance.

Now the Init*fixer becomes a copy of the item as well as its absolute
index within the list. This seems ugly, but for a user its very
difficult to obtain a copy of the current item in a different way,
and the use of the value seems to me like a justifying use case.
I.e: a tree list with per item padding depending of the level of each
node.

- changed dependent files accordingly but crude.