implemented crude relative focus movement
of the leaves through different exceptions to the FAN-OUT of the
Update-tree. Normally all messages become FAN-Out.
This means not changed while passing them on to all children,
so the leaves can react accordingly to the original message.
Now "alt+[left,right,up,down]" are altered immediately (in the root node)
, to result in the distribution of FocusLeave['s] which collect the path
to each leave.
The [focused] Leave now knows its own path and can emit a ChangeFocus
with a different (shorter) path to a different node, according to the
(relative) direction in which the focus should go.

For now we do not and can not know if the different path is valid.
So when the ChangeFocus has a wrong path, emit a error and a ChangeFocus
with a empty path which yields one default focus.

- renamed helper functions
- added a errList to model for (bad) debugging.
- added ids to nodePos also for debu.ging
- in boxer: reused FocusLeave msg, not only the path for better
  conserving of informations.